Seekfree_TC387_Opensource_Library0040

+++ code/pit/Pit.c
@@ -6,9 +6,9 @@
  */
 #include "Pit.h"
 
-const uint16 PIT00ms = 1;
-const uint16 PIT01ms = 5;
-const uint16 PIT10ms = 1;
+const uint32 PIT00ms = 1;
+const uint32 PIT01ms = 5;
+const uint32 PIT10ms = 1;
 
 void Pit_init(){
     pit_ms_init(CCU60_CH0, PIT00ms);
